Page bar upgrade (#570)
Update PageBar style, add Material ImageButton styles

I've reworked it as a ConstraintLayout, which works a lot nicer (page count is centred, the buttons on either side are chains which distribute themselves nicely in the available space).
Also added a dedicated button style, and added animations for when the buttons appear and disappear. Font and icon sizes have been reduced to material spec

The page counter and button areas are separated by some visual dividers (which I just noticed is what the Google keyboard does for suggestions) - this fixes the issue where the buttons didn't look centred due to the extra space around the page counter. Now you can clearly see where one space ends and another begins, so it looks balanced

I've added some explicit states to reflect what the page bar is displaying - basically whether it's a single page, or the first or last, or a normal middle page, or one where we don't know how many pages there are. This way it's clearer about defining what we need to show in each state, and how we show it.

I've also redone the code to show and hide the various elements - by doing the hiding first, things disappear cleanly and get replaced, instead of things appearing and overlapping before the other element disappears.