Fix login errors, improve "logged out" handling, add Authentication file
Moved "logged out" request error handling out of individual request callbacks -
now it's handled by default in AwfulFragment whenever a request ends with an
error. Fragments can tell their activity to re-authenticate (log out and show the
login activity), either returning to that activity or kicking the user back to
the main one.

Login status and reauth code has been moved to a new Authentication file, instead
of being internal to AwfulActivity. LogOutDialog is in there too - ideally this
is where you go to do anything related to the user's logged-in status.

SyncManager now queries Authentication to check if the user is logged in, and
aborts sync if not - avoids error messages when the user is on the login page.

Fixed an internal framework NPE from an early call to ViewPager#endDrag

Some cleanup and extra logging