Prevent view pager from jumping when swiping on a thread
Swiping on a code block in the thread view is meant to consume touch events,
allowing the user to scroll those blocks where possible. The view pager sees the
initial ACTION_DOWN event, but then the JavaScript handler sets a swipeEnabled
flag to false, forcing the view pager to ignore the rest of the events until
the flag is reset to true, when the JS handler sees the touch event end.

But if the thread is at the top or bottom, and the swipe ends up dragging the
page into a pull-to-refresh motion, the JS handler can lose focus and re-enable
the view pager in the middle of a swipe. The pager starts seeing MotionEvents
again, with MOVE events that are offset from the start of the drag, and it
interprets that as a sudden move and drags the pager by that amount.

I've fixed this by adding a latch to ignore the current move event, triggered
when swiping is disabled and reset when a new DOWN action is seen (representing
the start of a new touch event). So now all events are ignored when swiping is
disabled, and when it's enabled any current drag in progress will still be
ignored, until the next one starts.