Fix (hopefully) issue where the screen stays on by default for some people (#597)
Some people are reporting that the screen stays on in the thread view, like the
'keep screen on' menu option is selected (which is isn't), and they have to
toggle the option on and off again every time they start the app.

I can't replicate this problem, and it seems like it shouldn't happen anyway
(the flag is initialised to 'off' and requires the user to select the menu
option to change it), so I'm assuming it's a WebView issue. I've just made the
init code always explicitly set that behaviour to off when the WebView is first
created. Hopefully that's it!

Also removed some old API lint annotations