Fix launcher icon navigating the open app back to the forum index
NavigationEvent is treating the intent that the launcher sends when
opening the app as a "show the forum index" event. This means that when
the app is already open, tapping the icon in the launcher will resume
the app as normal in its saved state, but it will move the viewpager to
the forums index too.

This is happening because "show the forum index" is the default
NavigationEvent, so I've changed that to a new MainActivity event which
just means "show ForumsIndexActivity". If it's already created, it'll
just display in its current state without moving to a particular page.
If it needs to be created (e.g. on a cold open) the viewpager will
default to showing the forums index anyway.

NavigationEvent needs implementing fully anyway - probably best to have
the parsing function return null if a particular Intent isn't recognised
(i.e. no default) and define every event's Intent signature explicitly