Make pager and Forum and Thread fragments handle NavigationEvents
The fragments now receive the event types and handle them internally.
This means no other classes need to deconstruct things into parameters,
or have special methods for each navigation event (e.g. showThread)
with calls chained to the same method in the next component

The pager now takes a generic NavigationEvent and decides where it needs
to go - it's basically the last stop, so it decides to route "show forum"
events to the ForumDisplayFragment, etc. The main activity doesn't call
a "showForum" method on the pager - it doesn't need to make that decision,
it only needs to forward the appropriate events to the component that
handles them.

The downside here is that we have a NavigationEvent algebraic data type
where only some of those subtypes are actually handled by a method - so
instead of having a method for each type (which locks down exactly what
you can pass in) we have one method for the supertype, and some of those
subtypes will be invalid. So the method has to handle what it can, and
error on anything it can't - runtime failures instead of compile-time
basically. Not really any different from, say, a fragment expecting to
be able to cast its Activity to AwfulActivity, but still not ideal. It
would be nice if Kotlin let you define new ADTs from existing types
(like a HandledEvents type that is either NavigationEvent.Forum or
NavigationEvent.Bookmarks) so you could use that as your method parameter,
but it doesn't, sooo