Add NavigationEventHandler interface and implement in handlers
The interface is just a way of formalising handlers and giving a common
interface to the navigation system - basically just call #navigate and
it should pass on the event to whatever handles it

There's a problem though - a handler should handle a specific set of
events (e.g. ForumDisplayFragment handles Forum and Bookmarks events),
so it would be good to have an error fallback for when you pass an event
it doesn't handle (i.e. you've passed it to the wrong thing). BUT the
idea is also to be able to just call #navigate and let the event
propagate to where it's meant to go, so calling it from INSIDE the
fragment isn't an error, the fragment is just meant to say "I can't handle
this, so off it goes to the Activity" etc

The upshot of doing that is if an event is routed to the wrong destination,
that handler will pass it on as an unhandled event, and it'll eventually
come back and be passed on again... it'll cause a stack overflow in the end,
which is at least an obvious error... I'm just not sure if there's a better
way to handle this kind of general method with variable behaviour