Rework changelog to use HTML
Basically this makes the changelog look nicer, we can have formatting,
and we don't need to escape Every.Single.Apostrophe. which should make it
a lot easier to add Play Store release notes. The tradeoff is no auto
localisation, which... we're not using anyway

You can now supply a limit for the number of entries, so we can have
the whole changelog history in the app, but be selective about what's
shown. I've set it to show only the first (current) changelog entry on
app updates, and the settings menu lets you read the whole thing, if
you're somehow into that