Rework NavigationEventHandler interface
The interface now has two handler methods - one for the events it knows
about (and either consumes itself or passes along to a specific component)
and one routing events back up the hierarchy. So a fragment might handle
certain events, anything else it needs to pass to its parent - which in
turn may handle them or point them at another child component, or pass
them back up to the next level.

I'm not a fan of how this works out right now - ideally the #navigate
method should be public and final, since this is what everything should
call, and its internal logic just calls the other two methods. It should
never be overridden.

The two handler methods have default logic but can be overridden to add
a real implementation, but should only be called through #navigate. So
ideally they wouldn't be visible after definition. You could do this with
an object that implements the interface, but then you lose the ability to
override one of the methods in a subclass of the handler (e.g. giving
AwfulFragment a #defaultRoute implementation, then overriding #handleEvent
in a fragment that subclasses it, so you get both implementations)

Not sure there's a great way to enforce this but hey I'll have a look