Convert ForumsIndexActivity to kotlin and refactor Separated all the ImmersionMode stuff into its own activity - it's technically only used in ForumsIndexActivity, however the functionality is distinct, IMO, and cluttered up that class. It's cleaner to have it in its own activity. AwfulImmersionActivity extends from AwfulActivity, and ForumsIndexActivity extends from AwfulImmersionActivity. I plan to refactor the Immersion Mode stuff as well - it's messy and odd, and particularly the handler is prone to leaking and is difficult to parse what it's doing and when.
I also had to make a minor tweak to Intent.parse in NavigationEvent. As far as I can tell there is no way to call this from kotlin the same way there was from Java, so I made it a normal method that takes an intent as a parameter, instead of an extension method.