Removed PagerCallbacks, cleaned updateNavDrawer implementation
As far as I can tell, PagerCallbacks had no use. It only called setProgress (a deprecated method), which didn't show anything as there is already a custom progress bar implementation in the toolbar. I did pretty extensive testing but didn't find any bugs, so it appears to be a safe removal.