Convert ForumsIndexActivity to Kotlin, simplify Immersive, update libs (#610)
Reworked ForumsIndexActivity as a Kotlin class.

I got the immersion mode implementation down to about 5 lines.
I'm not sure why it was so over-complicated. Complications with previous version of android, perhaps? I tested it extensively and could find no issues with this implementation. I based it on the sample at https://github.com/googlesamples/android-ImmersiveMode so it should be kosher.

Also updated Gradle/Kotlin/all the outdated libraries