Fix loaders in ForumDisplayFragment
I've removed the code that destroys the loaders - that shouldn't need to
happen, just restarting them (which the code already did) should be
enough to refresh the data?

Also removed the ContentObserver stuff since it doesn't seem to actually
do anything to affect the loader, none of the callbacks are implemented
and the CursorLoader is meant to register its own ContentObserver
anyway.

I've tested this and the original working implementation, this seems to
have the same behaviour. Opening a thread with unread posts updates the
unread counter in the forum view, and so does turning the page in the
thread view. Refreshing the last page doesn't update the thread list
with a new last poster though. But these flaws are all in the original
version, so this fix hasn't introduced them at least!