Rework search filters, fix NavigationEvent.Url serialisation (#624)
The Url event wasn't fully implemented, so the #toIntent handling
basically created an Intent with no data. I've added that, so now
the Intents we create will parse back correctly.

I've overhauled search's filtering system - now there's a separate
SearchFilter class contains one of the filter types we handle, and
the data for that filter. The menu is automatically populated from
these types, each type can spawn a dialog for the data it needs, and
the fragment gets a callback when a new filter is added.

There are two ideas behind this. The first is implemented - the Search
NavigationEvent now takes an optional set of configured SearchFilters,
which means that other components can navigate to the Search section
and specify some filters to populate it with. I've added a "search
this forum" option to the thread list context menu which does this, but
you can specify any combination. "Search for this user in this thread",
"who quoted me today", that kind of thing.

The second idea is to make the filter UI easier to work with - ideally
something like a list of added filters, where you can edit each one
(maybe to change some terms) or swipe it away. Right now it still adds
each new filter to the text field like before, but they're represented
by SearchFilter objects - the fragment just needs to store them as a
data set and let the RecyclerView work with it. You can query each of
them to build up the final query string for the request, and keep that
text representation out of sight