Move bodyHtml into JS interface
The WebView needs to hold the body HTML for the current page, so we can
do things like comparing some new HTML and deciding whether to update
the page. The idea is to move that state out of the fragments containing
the WebView, so they can just pass in any new data and say "here, show
this" instead of being responsible for it

Since the actual AwfulWebView class is a widget it requires calls on the
UI thread, so it's easier to just hold the current data in the
WebViewJsInterface class instead along with the current preferences etc.
That way the page JS can call it for the current data - before each
fragment would subclass the interface class and wire the bodyHtml stuff
into itself