Skip body HTML updates when content hasn't changed
I've added a check to see if the HTML the app wants to display matches
the current HTML, i.e. there's been no change. This avoids "reloading"
the page contents when nothing is different, and stops the page position
from resetting

The HTML is held in the JS interface, and comparisons are done with
that version since the actual page DOM manipulates the content. We need
to compare with the base data, not how it currently looks

Need to test this to check it generally works. Explicit page reloads
with no changes should trigger this. A fully read page with a change
(e.g. an edited post) should ideally stay at the same scroll position,
even if the page was loaded with "jump to post" set (i.e. do it once).
New unread posts should probably just act as they do now - hide the old
ones (if enabled) and scroll to the first unread