Remove HTML state from WebView-using components
Now the WebView/JS Interface is handling HTML state internally, there's
no reason for the components using it to store the current HTML. They
can just pass it to the WebView for display when they get new content.

I've also removed WebViewClient#onPageFinished methods that were
refreshing the WebView with this stored state - as far as I can tell
they don't actually do anything. I assume the idea was to refresh the
page (and scroll position) when the content finishes loading and the
layout is stable. But that method only seems to fire once, when the
container HTML is loaded - after that we just switch out the contents
of the container div, and that doesn't trigger any loading callback.
Maybe we can add one?

Also removed the showBlankPage behaviour in ThreadDisplayFragment#syncPosts
since it's not really anything to do with syncing and all the sync calls
are preceded by a call to that method anyway