Fix loader behaviour with new Support Libraries (#630)
Updated support libraries to 27.1.1

Moved "current body HTML" state (i.e. the HTML that's displayed in the
container div) into the WebView components, instead of making the
WebView's owner responsible for storing it and providing it - now that
just pushes the data to the WebView when it wants something displayed.
The HTML state is stored in the WebViewJsInterface along with other
state like the preferences, and also avoids issues with being called on the
JS thread (since the WebView is a UI component on the main thread)

Added a check to skip body HTML updates when the content hasn't
changed. This fixes the "reloading" behaviour triggered by the loader that
was resetting the page scroll position - since the loader is just pulling the
same data from the cache (it's not even updating from the site) then the
contents haven't changed. An actual reload over the network will trigger
an update if that page data is different (edits, new posts etc). The new HTML
is compared to the last version passed to the WebView - *not* the HTML
currently being displayed, since the DOM may have changed e.g. rewriting
content, expanding or hiding posts etc.

Fixes #616

Some cleanup:

I've removed WebViewClient#onPageFinished methods that were
refreshing the WebView with this stored state - as far as I can tell
they don't actually do anything. I assume the idea was to refresh the
page (and scroll position) when the content finishes loading and the
layout is stable. But that method only seems to fire once, when the
container HTML is loaded - after that we just switch out the contents
of the container div, and that doesn't trigger any loading callback.
Maybe we can add one?

Also removed the showBlankPage behaviour in ThreadDisplayFragment#syncPosts
since it's not really anything to do with syncing and all the sync calls
are preceded by a call to that method anyway. 

Removed 'force'/'checkFirst' flags from html display functions -
not sure what this was even for, only one method call wanted to use this
'checkFirst' behaviour and I don't know why (it would ignore the request
to update the displayed html unless the page was blank, and I don't know
why you'd want to do that)