Fix jump-to-post behaviour for 'showthread.php' links
Fragment strings were being handled inconsistently, with the 'post'
prefix being stripped if you were redirected to a showthread.php link,
but not if you opened one directly. thread.js was readding the prefix,
so on some links you got 'postpost1234' and it obviously wouldn't work.

I moved all that code into the postJump setter so it formats everything
consistently and in one place. Now those URLs actually work I've
implemented the postJump parameter in the NavigationEvent's intent parser

Some apps helpfully strip out the fragment when passing a URL in as
Intent data (e.g. Contacts app), so there's not a lot we can do about that...