Implement Leper's Colony / Rap Sheets (#632)
This adds a Leper's Colony entry to the navigation menu, which opens a
browsing activity that requests and parses LC pages, generates a list of
Punishment rows, and uses a mustache template to display them. Opening it
and navigating between the LC and a user's rap sheet is done through a
NavigationEvent.LepersColony, which takes an optional user ID for
displaying an individual rap sheet instead of the main list

The display is basically repurposing the thread view system, including
its JS and CSS. The context menu basically acts like a post menu - the
mustache template adds the relevant attributes, and I've added a
conditional branch in thread.js that decides which menu to show, the one
for posts or the one for punishments.

Ideally we'd have things refactored so common webview stuff is in one
place, and things specific to say the thread view or the rap sheet view
could be mixed in as appropriate. That's a lot of work though, so right
now I'm just crowbarring this into the system that's built to display
threads (Announcements do this too, they're a lot closer to threads
though)

All the entry 'posts' are marked as unread so thread.js scrolls to the
first one when the HTML is changed. Not great but better than having it
stay halfway down the webview

Also added the site's probation/ban graphics, and created an "autoban"
version. I also switched the Announcements menu icon (from ic_error to
ic_announcement) while I was in there