Add more performant AwfulRequest class to speed up page parsing
When a request's response is handled, AwfulRequest uses jsoup to parse
the page data into a HTML Document, which involves building a tree of
nodes representing all the elements on the page. But the pages are fairly
complex structurally, and a big problem is the page-picker dropdowns -
they have an <option> element for every page in the thread/forum etc.
When this is a large number (in the thousands) each page parse can take
a very long time, several seconds between receiving the network response
and handling the document.

This PR uses a regex to strip out the page selector elements before the
HTML is parsed, resulting in a ~10x speedup when there are a significant
number of pages.

It basically works by adding a new abstract subclass of AwfulRequest,
called AwfulStrippedRequest. If you build a concrete request class from
this, there's a new handler method called #handleStrippedResponse, which
passes in a Document as usual (but with the page selectors removed), as
well as the current and last page numbers parsed from the page selector
(since this is important information and probably the only reason you'd
want the page selectors present in the Document anyway)

Right now only 3 requests implement it - the forum, thread and Leper's
Colony requests, which are the places you're likely to see large page
counts. Ideally every request would use it, but that means you need to
be aware that you aren't getting the full forum page. But we could work
something out, with e.g. opt-in flags if you *need* the entire structure,
so there's only one AwfulRequest base class

fixes #643