Remove hard failures when probed (#644)
I've removed the checks that cause response handlers to fail if the
user is on probation, because it stops pages from loading even when the
user should be able to use them.

For some reason, being on probation produces an AwfulError, even though
it's basically a state and not indicative of a site error etc. This is
recognised through an #isCritical check on AwfulError, which basically
whitelists PROBATION errors as being fine, and not something that should
halt response handling.

But then all the response handlers explicitly fail if #isCritical returns
true OR it's a probation. And I can't see any reason to do that, except
where probations actually block the user from using something, e.g.
composing a reply. And even then, it's better to handle it directly
instead of just letting the action fail