Convert AwfulRequest classes to Kotlin, rename some
I've converted all the implementations, partly to clean up and partly
to make future development a bit easier since we don't need to rely on
Java interop.

The forum index updater tasks got reworked a bit, including fixing a
crash in DropdownParser when a task failed and tried to show a Toast
on the wrong thread

PostRequest got renamed to ThreadPageRequest since it reflects what the
request actually handles better - it gets a page of posts, and also
processes thread data like the overall post count. Same for
ProfileRequest, renamed to RefreshUserProfileRequest since that's what
it actually does (and I removed the unused "username" parameter)

I've also renamed some Search requests for clarity, and also renamed the
"userTitle" preference and key to reflect the fact it's actually your
avatar's URL, *actually*

Add default implementation of handleError(), invert logic and reword docs

Most requests do exactly the same thing (swallow non-critical errors) so
it makes sense to have that as the default method implementation.

I've also flipped the logic so the method returns true if the error was
handled (which is the usual way of checking if an event has been consumed)