Defer navigation in ThreadDisplayFragment until attached to activity (#652)
It's possible to start a NavigationEvent that directs to the
ThreadDisplayFragment before that fragment has been created. This means
the fragment can be initialised by the viewpager, then immediately passed
the navigation event, before the fragment lifecycle has run.

Because opening threads etc. relies on access to the activity/context,
we need to defer handling this event until the fragment is attached to
an activity. So this commit basically checks if the fragment has been
added, and if not it stores the event, which is picked up and handled
in #onActivityCreated when it can be run safely.

I've moved the other state-restoration code in there too, and
consolidated some view restoration to avoid spreading things out across
all the lifecycle methods. Also changed some #setVisibility calls on the
FAB to #show and #hide since the linter was complaining that you can't
set it directly now.

Fixes #641