Fix blank initial pages when the WebView is (re)created
This basically does a refresh call when the HTML is loaded and the
container is first initialised. If the container wasn't ready when a
"display content" call came in, then it ends up loading it here

It's *possible* that some content might refresh the display twice, if
the container initialised refresh coincides with the app setting the
content, so there are two JS #showPageHtml calls after bodyHtml is set.
I'm only mentioning it in case it comes up as a "why's that happening"
issue - it should be rare if it ever happens, and it would only be a
single glitch when the WebView is first initialised

Fixes #653