Fix pull-to-refresh behaviour on non-full pages
P2R now refreshes if the currently displayed page is not full
(according to the user's posts-per-page setting). This avoids a situation
where the thread data updates, in the background, and the current page is
no longer the last one - previously this would trigger a page turn,
assuming the current page's posts are all being displayed since it's not
the last page, which wasn't the case when the displayed contents were
generated.

fixes #675