Fix loading behaviour when displaying the currently open thread
The last fix (ebb2d87) introduced an issue where #openThread would skip
the page load if the specified thread and page referred to the page
that was already being displayed. This was meant to allow the app to
navigate to the thread view without forcing a reload, but since thread
list entries open threads by specifying a thread ID and page, it wouldn't
reload to show new posts if that page was already open.

I've basically moved this check into #handleNavigation (since it's more
about navigating to the thread view and then deciding if anything else
needs to happen), and now #openThread always loads a page. I've reworked
the behaviour, so a NavigationEvent that only specifies a thread ID (and
no specific page number) will do nothing if that thread is what's
currently being displayed - meaning the NavigationEvent will effectively
just page to the ThreadDisplayFragment. If a different thread ID is
specified (or a certain page is requested) then a load will happen