Fix network request errors when probated or logged out
When I rewrote AwfulRequest I forgot a ! in the error-handling logic,
so everything broke when #checkPageErrors generated an AwfulError.

This actually caused two big issues (and probably more weird behaviour):

- non-critical errors (i.e. probations) were being treated as unhandled
critical errors and halting the whole request process with a failure
- not-logged-in errors (which you get when your cookie expires) were
being treated as handled errors and basically acting as a successful
request even though they'd failed

So I fixed that logic - now critical errors fail unless handled, and
non-critical ones are ignored. I've also fixed up a couple of request
classes that weren't ignoring non-critical errors (so they didn't work
while on probation)

I've also fixed the cookie expiry logic in
CookieController#restoreLoginCookies which is also used as a check for
if the user is logged in - the validity check wasn't looking to see if
the cookie had expired, so it could say you were logged in but actually
end up with no cookies in use (because they'd all expired). This wasn't
actually the problem - loading a page without any cookies should return
a page with the "not logged in" message that generates an AwfulError and
kicks you to the login dialog, but the above issue with critical errors
being swallowed was stopping that from happening. But this logic is more
correct anyway.

Fixes #684, fixes #596