Remove non-critical AwfulErrors from error handling
We had a situation where our error-handling code (and anything
overriding it) had to do a check for non-critical errors and return a
"handled" response. Nothing ever cared about doing anything with these
(which at this moment are only "you're probated" AwfulErrors), so I've
removed all that boilerplate handling code and made the main request
handler code do a check, so it only passes critical errors for handling.

This means individual requests can't handle non-critical errors in any
special way anymore, but we only have one kind and there's no reason to
handle it anyway. But if that changes in the future (or we add new and
exciting non-critical errors) this might need to be rolled back