Fix blank OLED views, excitable post previews and separator lines
Post preview and private message webviews have the appropriate CSS
classes applied to their wrapper HTML, so content is actually styled
as a post and is visible on all themes including OLED!

I've added a div to act as a separator between the post header and the
post content, instead of having one of those add a border. Some themes
defined it in the post content styling, meaning you'd get a weird line
appearing at the top of the preview when there's nothing to separate.

Now the header and content are styled as isolated elements, and they
just need to define their vertical padding to act as spacing from any
separator. Makes the separator more flexible too, if you want to do
something weird with it...

And there was some issue where the dialog popup would keep re-laying out
(maybe more obvious when network delays give the loading spinner time to
be visible and then disappear), and half the time the end result would
have wildly different heights for the same post content.

I've switched to a ConstraintLayout and removed the NestedScrollView
(the thread viewer doesn't use one either), and also swapped to a
standard DialogFragment instead of the old legacy AwfulDialogFragment
which might be causing its own issues. It seems to be behaving itself
a lot better now - it even looks nicer since the content pops into view
and smoothly resizes the popup