Fix issues with post menu (...) not working
I've reverted the newer CSS which seems to be creating compatibility
issues, and has inconsistent behaviour over multiple devices and
emulators.

The issue is that the menu responds to a click listener on
the .postmenu element, but the actual visual icon to tap is a pseudo
element - some webview versions won't treat this as the element with
the listener, and rely on that element being in the same place (so it
gets touched by the user). But when it's styled with display:contents
it can be on another part of the screen, or completely collapsed - so
the click target isn't under the ... icon anymore.

Just to make life easier I've redone the CSS so the element and the
pseudo element are both in the same 24x24px space with basic padding
and margins, so hopefully it will stop the issues and people constantly
asking what the deal is