Fix post menu CSS even better
Some larger text settings would let the ... glyph expand to the point
where it pushed out of the defined touch area, making the page wider
and introducing a horizontal scrollbar.

I've made that glyph a fixed size since it's basically an icon and
shouldn't be changing with the page's font size anyway, and adjusted
the CSS to position it better and give it a 40px touch target, which
basically sits in the top right of the post header. It's not exactly
where it was aligned before, but it's in the right place for a 40px
button.

I set a font size that matches what I was already seeing with my current
(small) text size - it's smaller than the ... icon in the toolbar, which
I think looks better and clearer than matching it. I also tweaked the
margins slightly so that themes with full-width post headers (OLED and
Classic) have the post menu icon line up horizontally with the toolbar
one - before they were sliiiightly off and that looks pretty bad.
The other themes have a clear visual offset so they don't look like
the icons should align.

(if you do want to match the toolbar icon size, 18pt seems about right)