Update to new version code scheme
Version codes now follow a MmmRR pattern, major number followed by two
digits for minor and revision respectively. This means 100 revisions for
each minor release, and 100 minors for each major - if we ever need more
than this (!) we'll have to extend the number of digits, but since the max
version code is 2,100,000,000 I didn't want to push it by jumping
straight to 3,000,000. Sure hope this ends up not being a bad decision!

So the basic idea now (explained in the manifest) is that we push betas
straight to production when ready. The next beta increments the minor
number. Updates to this beta and fixes for production both increment
their revision number - the beta being one minor number higher means
production can be updated without surpassing the beta version. Once the
current beta revision is ready to roll out, we just push it to production,
merge beta into master and tag it