Add WebViewSearchBar for your web view searching needs
Just a basic widget for replacing the broken implementation in the
WebView itself. As a bonus it's themed to match the toolbar now!

I've added the counter that shows how many results you have/which one is
currently highlighted, and tried to style it like Chrome does it (on the
desktop anyway) - so basically the count area can expand and push back on
the typey bit, but the longer the part you're typing the less likely you
are to get a lot of results, so it shouldn't be an issue

Also when I came to use the carets for the prev/next buttons, I
noticed the down one was bigger - different viewport settings for
whatever reason, and I couldn't change those on the up caret without
getting a weird offset. So I've switched out references to that big caret
for the smaller one, and adjusted the scaling on those image views to
compensate

That's the subforum expander and the down button to show the original
PM message - the former should look about the same, I can't expand the PM
button image without the button growing with it, so I left it but it looks
fine. I've also deleted an XML thing that wasn't being used (pretty sure
it was abandoned in favour of the spinny expander arrows)