CaptchaActivity: new activity for prompting the user with captchas
This activity displays the Cloudflare captchas on SA to the user in a
web view. If the user manages to solve the captcha, the resulting
captcha clearance cookie is parsed and stored in the CookieController
used for AwfulRequest instances.

The class exposes a static member to check whether a captcha activity
is currently active. This is necessarily to constrain some
behaviours (such as displaying the login activity) while a challenge
is in progress.