AwfulRequest: set user-agent from captcha challenge if set
Cloudflare only validates a captcha clearance token if the user-agent
matches, however by default Volley has a different user-agent from the
web view in which the captcha is solved.

To work around this, CaptchaActivity persists the last user-agent it
used for a captcha challenge and this commit updates AwfulRequest to
use that username if it is present.