CaptchaActivity: add handler to launch captcha challenge on errors
Adds a method which is intended to be called in implementations of
AwfulResultCallback.failure. If an activity is present and the error
indicates a captcha challenge, the captcha activity is launched and
presented to the user.

The method is called in AwfulLoginActivity, ForumDisplayFragment and
ThreadDisplayFragment where appropriate, which should remove all
situations where a user can get stuck without the ability to trigger
the captcha challenge.

In general it would be nice if we could do this in a centralised
place, but AwfulRequest does not have the activity from which it was
initiated, so it can't happen in there. There are likely many more
places where this would be appropriate to add, but I simply could not
be bothered to do that yet.